requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2
neuralprophet==0.9.4
prophet==1.1.6
causalimpact==0.2.4
xgboost==2.1.2
shap==0.46.0
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
python-jose[cryptography]==3.3.0
psycopg[binary]==3.2.3
httpx==0.27.2
